fix: ignore DOM node which display property is none 1. move const UUID_ATTR to appendUuid.ts 2. move logic to class RenderTree
